backend/requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
defusedxml==0.7.1
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.4
fpdf2==2.8.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
PyJWT==2.10.1
requests==2.32.4
urllib3==2.5.0
Werkzeug==3.1.3

# CRITICAL MISSING DEPENDENCY - Required for OCR functionality
google-cloud-vision==3.4.5

# RECOMMENDED ADDITIONS for enterprise scale and security
python-dotenv==1.0.0
psycopg2-binary==2.9.9
bcrypt==4.1.2

gunicorn==20.1.0 
